% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\name{autocovariance_test}
\alias{autocovariance_test}
\title{Test for equality of autocovariance functions for two (linearly Weighted_Fixed_Lag) stationary time series}
\usage{
autocovariance_test(
  X,
  Y,
  max_lag = NULL,
  test = "Auto_Lag_Bartlett",
  trunc = NULL,
  num_bootstrap = 500,
  block_size = NULL,
  prewhiten = TRUE,
  dependent_series = TRUE,
  bootstrap_fixed_stats = FALSE,
  weights = NULL,
  average_bartlett_cov = TRUE
)
}
\arguments{
\item{X}{a \eqn{n x m} mean zero (column-wise) stationary time series with \eqn{m < n}. Must be a matrix.}

\item{Y}{a \eqn{n x m} mean zero (column-wise) stationary time series with \eqn{m < n}. Must be a matrix.}

\item{max_lag}{the maximum lag to be considered. Must be a positive integer less than \eqn{n}. Note that the automatic lag selection tests may choose \code{max_lag} less than the one supplied. If not supplied, \code{max_lag = ceiling((log2(n))^0.9999)}.}

\item{test}{the tests to be performed. Must be a vector containing a subset of \code{"Fixed_Lag"}, \code{"Weighted_Fixed_Lag"}, \code{"Auto_Lag_Jin"} and \code{"Auto_Lag_Bartlett"}.}

\item{trunc}{for the \code{"Fixed_Lag"}, \code{"Weighted_Fixed_Lag"} and \code{"Auto_Lag_Bartlett"} tests, the truncation rule used in Bartlett's formula. If not supplied, \code{trunc = floor(n^(1/3))}.}

\item{num_bootstrap}{for the \code{"Auto_Lag_Jin"} and \code{"Auto_Lag_Bartlett"} tests or the \code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests with \code{bootstrap_fixed_stats = TRUE}, the number of bootstrap resamples to be used in the Block of Blocks algorithm. Must be a positive integer.}

\item{block_size}{for the \code{"Auto_Lag_Jin"} and \code{"Auto_Lag_Bartlett"} tests or the \code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests with \code{bootstrap_fixed_stats = TRUE}, the block length to be used in the Block of Blocks algorithm. Must be a positive integer less than \eqn{n}. If not supplied, \code{block_size = max(floor(0.5 * n^(1/3)), 2)}.}

\item{prewhiten}{for the \code{"Auto_Lag_Jin"} and \code{"Auto_Lag_Bartlett"} tests or the \code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests with \code{bootstrap_fixed_stats = TRUE}, should the supplied time series be prewhitened? \code{prewhiten = TRUE} is strongly recommended.}

\item{dependent_series}{for the \code{"Fixed_Lag"}, \code{"Weighted_Fixed_Lag"} and \code{"Auto_Lag_Bartlett"} tests, should the series be assumed dependent?}

\item{bootstrap_fixed_stats}{for the \code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests, should the test null distribution of the statistics be bootstrapped?}

\item{weights}{for the \code{"Weighted_Fixed_Lag"} test, the weight placed on each lag. Should be a vector of length \code{max_lag + 1} with values between \eqn{0} and \eqn{1}. By default, weights are \code{1, max_lag/(max_lag + 1), \dots, 1/(max_lag + 1)}.}

\item{average_bartlett_cov}{for the \code{"Auto_Lag_Bartlett"} test, should the covariance be estimated on the original sample or be averaged over each bootstrap resample?}
}
\value{
A named list containing many relevant statistics pertaining to each test
\itemize{
\item \code{delta} - the sample difference in autocovariance functions.
\item \code{dep_cov} - the asymptotic covariance matrix of the autocovariance differences computed assuming dependence of the series.
\item \code{ind_cov} - - the asymptotic covariance matrix of the autocovariance differences computed assuming independence of the series.
\item \code{fixed_lag_stat} - for the \code{"Fixed_Lag"} test, the test statistic.
\item \code{fixed_lag_df} - for the \code{"Fixed_Lag"} test, the degrees of freedom associated with the test statistic.
\item \code{fixed_lag_weight_stat} - for the \code{"Weighted_Fixed_Lag"} test, the weighted test statistic.
\item \code{fixed_lag_weight_alpha} - for the \code{"Weighted_Fixed_Lag"} test, the scale parameter associated with the weighted test statistic.
\item \code{fixed_lag_weight_beta} - for the \code{"Weighted_Fixed_Lag"} test, the shape parameter associated with the weighted test statistic.
\item \code{fixed_lag_pval} - for the \code{"Fixed_Lag"} test, the associated p-value.
\item \code{fixed_lag_weight_pval} - for the \code{"Weighted_Fixed_Lag"} test, the associated p-value of the weighted test.
\item \code{auto_lag_jin_stat} - for the \code{"Auto_Lag_Jin"} test, the test statistic.
\item \code{auto_lag_bart_stat} - for the \code{"Auto_Lag_Bartlett"} test, the test statistic.
\item \code{auto_lag_jin_L} - for the \code{"Auto_Lag_Jin"} test, the automatically selected lag.
\item \code{auto_lag_bart_L} - for the \code{"Auto_Lag_Bartlett"} test, the automatically selected lag.
\item \code{auto_lag_jin_cov} - for the \code{"Auto_Lag_Jin"} test, the bootstrapped asymptotic covariance matrix.
\item \code{auto_lag_bart_cov} - for the \code{"Auto_Lag_Bartlett"} test, the bootstrapped asymptotic covariance matrix.
\item \code{auto_lag_jin_pval} - for the \code{"Auto_Lag_Jin"} test, the pseudo p-value.
\item \code{auto_lag_bart_pval} - for the \code{"Auto_Lag_Bartlett"} test, the pseudo p-value.
\item \code{dependent_series} - the provided \code{dependent_series} argument.
\item \code{bootstrap_fixed_stats} - the provided \code{bootstrap_fixed_stats} argument.
\item \code{max_lag} - the provided \code{max_lag} argument.
\item \code{k} - the time series dimension.
\item \code{n} - the time series length.
\item \code{var_names} - the shared column names of \code{X} and \code{Y}.
}
}
\description{
Perform a hypothesis test for equality of autocovariance functions for two time series with one or more of the following methods: (Weighted) Fixed_Lag, (Weighted) Weighted_Fixed_Lag, Bootstrapped Weighted_Fixed_Lag and Bootstrapped Bartlett. The former two tests assess equality up to a fixed lag, while the latter two select the "optimal lag" for testing the hypothesis using an AIC-like penalty at each lag. The tests can handle multivariate time series, but the computations become considerably more intense with an increase in dimension.
}
\details{
Consider two \eqn{m}-dimensional, stationary time series with mean zero. \code{autocovariance_test} tests for equality of autocovariance functions of the respective series. The independent \code{"Fixed_Lag"} test is given by Lund et. al (2009). Their test assumes
independence of the two series and quantifies the distribution of the autocovariance differences through Bartlett's formula (see Brockwell and Davids (1991)). \code{autocovariance_test} provides the corresponding asymptotic covariance matrix under the null, \code{ind_cov},
the chi-square test statistic, \code{fixed_lag_stat} and its associated degrees of freedom and p-value, \code{fixed_lag_df} and \code{fixed_lag_pval}. An extension of the \code{"Fixed_Lag"} test to linearly dependent series can be run by setting \code{dependent_series = TRUE}.
If the independence of the series is not obvious, the \code{dependent_series = TRUE} test can be used without penalty. Both tests assess equality of autocovariances up to a fixed lag \code{max_lag}. The \code{trunc} parameter controls the truncation of the Bartlett formula infinite
sum (see the vignette for more details). Any truncation rule must follow the assumptions of Theorem A.1 in Berkes et al. (2006).

\code{autocovariance_test} also computes weighted variants of the \code{"Fixed_Lag"} tests using \code{test = "Weighted_Fixed_Lag"}. The weighting idea is borrowed from goodness-of-fit testing (see Fisher and Gallagher (2012)). By default, we assign to lags \eqn{0, 1, \dots , L} weights
\eqn{1, max_lag/(max_lag + 1), \dots , 1/(max_lag + 1)}. This can be controlled via the \code{weights} argument. The corresponding test statistics have distributions that can be approximated by a Gamma distribution. For the \code{"Weighted_Fixed_Lag"} test, \code{autocovariance_test} outputs the test statistic, \code{fixed_lag_weight_stat}, its gamma distribtion
parameters, \code{weighted_fixed_lag_alpha} and \code{weighted_fixed_lag_beta}, and the associated p-value \code{weighted_fixed_lag_pval}. All previously mentioned tests require Gaussianity of the series. The following two tests can be applied to series with non-normal distributions.

The \code{"Auto_Lag_Jin"} (see Jin et. al (2019)) and \code{"Auto_Lag_Bartlett"} tests compute a similar test statistic to the \code{"Fixed_Lag"} test but through a block of blocks bootstrap. The original test statistic is monotone increasing with lag, so the tests apply
an AIC-like penalty to choose the "best" lag (up to \code{max_lag}) to perform the hypothesis test with. This provides some desirable properties under the alternative hypothesis. \code{prewhiten = TRUE} fits a large VAR(p) process to \code{X} and \code{Y} and then applies
the tests to the residuals of said process. It is highly recommended as prewhitening improves empirical rejection rates. \code{num_bootstrap} and \code{block_size} control the number of bootstrap resamples and block length used in the \code{"Auto_Lag_Jin"} and \code{"Auto_Lag_Bartlett"} tests. \code{"Auto_Lag_Bartlett"} replaces the Jin
et. al (2019) covariance estimate with the \code{"Fixed_Lag"} estimate averaged over each bootstrap resample. For the \code{"Auto_Lag_Jin"} test, \code{autocovariance_test} outputs the test statistic, \code{auto_lag_jin_stat}, the optimally selected lag, \code{auto_lag_jin_L}, the
estimated covariance under the null, \code{auto_lag_jin_cov}, and psuedo p-value \code{auto_lag_jin_pval}. The same statistics for \code{"Auto_Lag_Bartlett"} are given by \code{auto_lag_bart_stat}, \code{auto_lag_bart_L}, \code{auto_lag_bart_cov} and \code{auto_lag_bart_pval}. Alternatively, one may choose to bootstrap the
\code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests by setting \code{bootstrap_fixed_stats = TRUE}. This uses the same covariance estimate as the \code{"Fixed_Lag"} and \code{"Weighted_Fixed_Lag"} tests, but approximates the null distribution via the Block of Blocks bootstrap.

For more on key assumptions behind each of the tests and a thorough example, please see the vignette.
}
\examples{
## Simulate two marignal AR(1) processes that are correlated at lag zero and run tests
set.seed(12345)

# Simulate correlated errors
Sigma <- matrix(c(1, 0.8, 0.8, 1), 2, 2) # covariance of errors
n <- 1000 # sample size
Errors <- mvtnorm::rmvnorm(n, mean = c(0, 0), sigma = Sigma)

# Simultate correlated AR(1) processes under the null hypothesis
X <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 1])) 
Y <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 2]))

# Run tests for equality of autocovariance functions up to lag 5
output <- autocovariance_test(X, Y, max_lag = 5, 
test = c("Fixed_Lag","Weighted_Fixed_Lag","Auto_Lag_Jin","Auto_Lag_Bartlett"))
# All tests fail to reject the null
output

## In males and females, do monthly lung deaths in the UK from 1974-1979
## have the same autocovariance structure?
## Here we carefully go through the process of making sure the series 
## adhere to the test assumptions
?UKLungDeaths 

n <- length(mdeaths) # series length
# plot series
plot(mdeaths)
plot(fdeaths)
# Obvious non-zero means and not stationary 

# Standardize by monthly means and standard deviations to ensure stationarity
male_matrix <- matrix(mdeaths, 6, 12, byrow = TRUE)
female_matrix <- matrix(fdeaths, 6, 12, byrow = TRUE)
# standardize series
male_stnd <- c(t(apply(male_matrix, 2, scale)))
female_stnd <- c(t(apply(female_matrix, 2, scale)))

# Plot again to check stationarity
t <- 1:n
plot(t, male_stnd, type = "l") # definitely has a trend
plot(t, female_stnd, type = "l") # maybe a slight trend

# Detrend both series to adhere to mean zero assumption + stationarity
male_stnd <- lm(male_stnd ~ t)$residuals
female_stnd <- lm(female_stnd ~ t)$residuals
plot(t, male_stnd, type = "l") # reasonably stationary, 
plot(t, female_stnd, type = "l") # reasonably stationary

# Both series are reasonably Gaussian
qqnorm(male_stnd)
qqline(male_stnd)
qqnorm(female_stnd)
qqline(female_stnd)

# Check for correlation
ccf(female_stnd, male_stnd) # Clear lag zero correlation

# The series are likely gaussian and Weighted_Fixed_Lag, hence it 
# makes most sense to use "Weighted_Fixed_Lag" test
output2 <- autocovariance_test(matrix(male_stnd), matrix(female_stnd), 5, 
test = "Weighted_Fixed_Lag")
output2

# Both the weighted and unweighted tests fail to reject the null hypothesis
# The weighted test comes close to rejecting, which makes sense given the large difference in lag 0
acf(cbind(male_stnd, female_stnd), type = "covariance", lag.max = 5)

}
\references{
Berkes, I., Horvath, L., Kokoszka, P. and Shao, Q. M. (2006). \emph{On discriminating between longrange dependence and changes in the mean. Annals of Statistics, 34, 1140–65}\cr
\cr Brockwell, P. J., and Davis, R. A. (1991). \emph{Time series: theory and methods (Second ed.). New York: Springer-Verlag. }\cr
\cr Fisher, T. J., and Gallagher, C. M. (2012).  \emph{New weighted portmanteau statistics for time series goodness  of  fit  testing, J.  Amer.  Statist.  Assoc., 107(498),  777–787}\cr
\cr Jin, L., Cai, L., and Wang, S. (2019).  \emph{A computational bootstrap procedure to compare two Weighted_Fixed_Lag  time  series, Journal  of  Statistical  Computation  and  Simulation, 89(15), 2831–2847}\cr
\cr Lund, R., Bassily, H., and Vidakovic, B. (2009). \emph{Testing equality of stationary autocovariances, Journal of Time Series Analysis, 30(3), 332–348}\cr
}
