\name{autocovarianceTesting-package}
\alias{autocovarianceTesting-package}
\alias{autocovarianceTesting}
\docType{package}
\title{
\packageTitle{autocovarianceTesting}
}
\description{
\packageDescription{autocovarianceTesting}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{autocovarianceTesting}
\packageIndices{autocovarianceTesting}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{autocovarianceTesting}

Maintainer: \packageMaintainer{autocovarianceTesting}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
