% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calculateAutocovariance}
\alias{calculateAutocovariance}
\title{Compute computes the DEMEANED autocovariance function of two concatenated multivariate time series}
\usage{
calculateAutocovariance(X, Y)
}
\arguments{
\item{X}{a time series}

\item{Y}{a time series}
}
\description{
Compute computes the DEMEANED autocovariance function of two concatenated multivariate time series
}
