% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionsUser.R
\name{plot.acvfTest}
\alias{plot.acvfTest}
\title{Plot Difference of Autocovariance Functions}
\usage{
\method{plot}{acvfTest}(x, bar_color = "gray20", var_names = NULL, include_ci = FALSE)
}
\arguments{
\item{x}{acvfTest object, result of autocovariance_test}

\item{bar_color}{The color for the bars in the plot of differences, defaults at \code{"gray10"}.}

\item{var_names}{An optional vector of length matching the dimension of the time series. Only used when greater than one to label the grid plot.}

\item{include_ci}{A flag to indicate whether to include 95\\% Simultaneous Confidence Intervals on the differences of autocovariances. See details.}
}
\value{
An unnamed list containing base R graphic plot elements is supplied.
}
\description{
Will built a grid of \eqn{m x m} plotting the difference in the autocovariance function of two \eqn{m}-dimensional time series based on the output of autocovariance_test()
}
\details{
Confidence bands are plotted when \code{include_ci=TRUE} based on 95\\% simultaneous confidence intervals.

Specifically, a vector \eqn{\boldsymbol{\Delta}} is calculated as the difference of the autocovariance function between two series. Based on Bartlett's formula we calculate a covariance matrix for \eqn{\boldsymbol{\Delta}}, called \eqn{\mathbf{W}}.
The quadratic form \eqn{n\boldsymbol{\Delta}'\mathbf{W}^{-1}\boldsymbol{\Delta}} is asymptotically a Chi-square random variable. 95\\% critical values are calculated from the corresponding chi square distribution along with marginal variance terms from \code{diag}\eqn{(\mathbf{W})} to construct a 95\\% simultanous confidence interval for each autocovarince difference in \eqn{\boldsymbol{\Delta}}.

It is worth noting the multivariate nature of the proposed method. It is possible to have a significant result in \code{autocovariance_test} but for all differences to be within their respective simultaneous intervals.
}
