% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionsUser.R
\name{plot_autocovariances}
\alias{plot_autocovariances}
\title{Construct a plot of the Autocovariance Function for two time series}
\usage{
plot_autocovariances(
  X,
  Y,
  max_lag = NULL,
  ggplot2 = FALSE,
  plot_options = list(series_names = c("Series 1", "Series 2"), bar_colors =
    c("gray10", "gray50"), var_names = NULL)
)
}
\arguments{
\item{X}{a \eqn{n x m} time series with \eqn{m < n}. Must be a matrix, vector or ts object.}

\item{Y}{a \eqn{n x m} time series with \eqn{m < n}. Must be a matrix, vector or ts object.}

\item{max_lag}{the maximum lag to be considered. Must be a positive integer less than \eqn{n}.  If not supplied, \code{max_lag = ceiling((log2(n))^0.9999)}.}

\item{ggplot2}{a flag on whether to construct a ggplot2 version of the plot.}

\item{plot_options}{a list containing elements to customize the plot, see details.}
}
\value{
If \code{ggplot2=FALSE}, an unnamed list containing base R graphic plot elements is supplied. If \code{ggplot2=TRUE}, a \code{gg} or \code{ggplot} object is returned.
}
\description{
A convenience function that will construct a plot of the sample Autocovariance function for two (potentially multivariate) time series. The function will be a $k x k$ grid of side-by-side bar graphs showing the autocovariance value at each lag for each series.
}
\details{
The parameter \code{plot_options} is a list that contains three elements. Defaults are supplied. If the user attempts to change them error checking is performed and the user may receive a warning.
\itemize{
\item \code{series_name} - a vector of two strings specifying names for the two series, defaults as "Series 1" and "Series 2"
\item \code{bar_colors} - a vector of two strings specifying the colors for the bars of the autocovariance values, defaults as "gray10" and "gray50". The code also checks to make sure legitimate color choices are made.
\item \code{var_names} - an optional vector of strings with length matching the dimensionality of the series specifying the names of the time series within the multivariate series. If not supplied, or incorrectly supplied, defaults to the convention Var1, Var2, ...
}
}
\examples{
library(ggplot2)

## Simulate two marignal AR(1) processes that are correlated at lag zero and run tests
set.seed(12345)

# Simulate correlated errors
Sigma <- matrix(c(1, 0.8, 0.8, 1), 2, 2) # covariance of errors
n <- 1000 # sample size
Errors <- mvtnorm::rmvnorm(n, mean = c(0, 0), sigma = Sigma)

# Simultate correlated AR(1) processes under the null hypothesis
X <- arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 1]) 
Y <- arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 2])

## Base R Graphics
## Default options
plot_autocovariances(X, Y)

## Plot options
plot_autocovariances(X, Y, max_lag=7,
                     plot_options=list(series_names = c("My AR1", "Your AR1"),
                                       bar_colors = c("forestgreen", "gray20")))
                                       
## ggplot2 Graphics
## Default options
plot_autocovariances(X, Y, ggplot2=TRUE)

## Plot options
plot_autocovariances(X, Y, max_lag=7, ggplot2=TRUE,
                     plot_options=list(series_names = c("My AR1", "Your AR1"),
                                       bar_colors = c("forestgreen", "gray20")))
                                     
#########
## A mulitviariate example
set.seed(12345)
Sigma11 <- matrix(c(1, 0.6, 0.6, 1), 2, 2) # covariance of errors
Sigma12 <- matrix(c(0.1, 0.2, 0.2, 0.1), 2, 2) # Cross-covariance
Sigma <- rbind(cbind(Sigma11,Sigma12), cbind(Sigma12,Sigma11))
n <- 1000 # sample size
Errors <- mvtnorm::rmvnorm(n, mean = c(0, 0, 0, 0), sigma = Sigma)

# Simultate correlated AR(1) processes under the null hypothesis
X1 <- arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 1]) 
X2 <- arima.sim(model = list(ar  = 0.5), n, innov = Errors[ , 2]) 
Y1 <- arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 3])
Y2 <- arima.sim(model = list(ar  = 0.5), n, innov = Errors[ , 4])
X <- cbind(X1, X2)
Y <- cbind(Y1, Y2)

plot_autocovariances(X, Y, max_lag=7,
                     plot_options=list(series_names = c("My VAR1", "Your VAR1"),
                                       bar_colors = c("forestgreen", "gray20"),
                                       var_names = c("Dim One", "Dim 2")))

plot_autocovariances(X, Y, max_lag=7, ggplot2=TRUE,
                     plot_options=list(series_names = c("My VAR1", "Your VAR1"),
                                       bar_colors = c("forestgreen", "gray20"),
                                       var_names = c("Dim One", "Dim 2")))
}
