% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\name{print.acvfTest}
\alias{print.acvfTest}
\title{Summarizing autocovarianceTest function output}
\usage{
\method{print}{acvfTest}(x, ...)
}
\arguments{
\item{x}{a \code{"acvfTest"} object given by \code{"autocovarianceTest()"}.}

\item{...}{additional arguments to \link[base]{print}.}
}
\description{
Summarizing autocovarianceTest function output
}
\examples{
#' ## Simulate two marignal AR(1) processes that are correlated at lag zero and run tests
set.seed(12345)

# Simulate correlated errors
Sigma <- matrix(c(1, 0.8, 0.8, 1), 2, 2) # covariance of errors
n <- 1000 # sample size
Errors <- mvtnorm::rmvnorm(n, mean = c(0, 0), sigma = Sigma)

# Simultate correlated AR(1) processes under the null hypothesis
X <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 1])) 
Y <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 2]))

# Run tests for equality of autocovariance functions up to lag 5
output <- autocovarianceTest(X, Y, L = 5, 
test = c("Independent","Dependent","bootDependent","bootBartlett"))
# All tests fail to reject the null
print(output)
}
