% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\name{autocovarianceTest}
\alias{autocovarianceTest}
\title{Test for equality of autocovariance functions for two (linearly dependent) stationary time series}
\usage{
autocovarianceTest(
  X,
  Y,
  L = NULL,
  test = "Dependent",
  B = 500,
  b = NULL,
  prewhiten = TRUE,
  plot = TRUE
)
}
\arguments{
\item{X}{a \eqn{n x m} mean zero (column-wise) stationary time series with \eqn{m < n}. Must be a matrix.}

\item{Y}{a \eqn{n x m} mean zero (column-wise) stationary time series with \eqn{m < n}. Must be a matrix.}

\item{L}{the maximum lag to be considered. Must be a positive integer less than \eqn{n}. Note that the automatic lag selection tests may choose \eqn{L} less than the one supplied. If not supplied, \code{L = ceiling((log2(n))^0.9999)}.}

\item{test}{the tests to be performed. Must be a vector containing a subset of \code{"Independent"}, \code{"Dependent"}, \code{"bootDependent"} and \code{"bootBartlett"}.}

\item{B}{for the \code{"bootDependent"} and \code{"bootBartlett"} tests, the number of bootstrap resamples to be used in the Block of Blocks algorithm. Must be a positive integer.}

\item{b}{for the \code{"bootDependent"} and \code{"bootBartlett"} tests, the block length to be used in the Block of Blocks algorithm. Must be a positive integer less than \eqn{n}. If not supplied, \code{b = max(floor(0.5 * n^(1/3)), 2)}.}

\item{prewhiten}{for the \code{"bootDependent"} and \code{"bootBartlett"} tests, should the supplied time series be prewhitened? \code{prewhiten = TRUE} is strongly recommended.}

\item{plot}{should a plot of the tested sample autocovariances be made?}
}
\value{
A named list containing many relevant statistics pertaining to each test
\itemize{
\item \code{delta} - for the \code{"Independent"} and \code{"Dependent"} tests, the sample difference in autocovariance functions.
\item \code{dep_cov} - for the \code{"Dependent"} test, the asymptotic covariance matrix of the autocovariance differences.
\item \code{ind_cov} - for the \code{"Independent"} test, the asymptotic covariance matrix of the autocovariance differences.
\item \code{ind_stat} - for the \code{"Independent"} test, the test statistic.
\item \code{ind_df} - for the \code{"Independent"} test, the degrees of freedom associated with the test statistic.
\item \code{ind_weight_stat} - for the \code{"Independent"} test, the weighted test statistic.
\item \code{ind_alpha} - for the \code{"Independent"} test, the scale parameter associated with the weighted test statistic.
\item \code{ind_beta} - for the \code{"Independent"} test, the shape parameter associated with the weighted test statistic.
\item \code{ind_pval} - for the \code{"Independent"} test, the associated p-value.
\item \code{ind_weight_pval} - for the \code{"Independent"} test, the associated p-value of the weighted test.
\item \code{dep_stat} - for the \code{"Dependent"} test, the test statistic.
\item \code{dep_df} - for the \code{"Dependent"} test, the degrees of freedom associated with the test statistic.
\item \code{dep_weight_stat} - for the \code{"Dependent"} test, the weighted test statistic.
\item \code{dep_alpha} - for the \code{"Dependent"} test, the scale parameter associated with the weighted test statistic.
\item \code{dep_beta} - for the \code{"Dependent"} test, the shape parameter associated with the weighted test statistic.
\item \code{dep_pval} - for the \code{"Dependent"} test, the associated p-value.
\item \code{dep_weight_pval} - for the \code{"Dependent"} test, the associated p-value of the weighted test.
\item \code{jin_stat} - for the \code{"bootDependent"} test, the test statistic.
\item \code{bart_stat} - for the \code{"bootBartlett"} test, the test statistic.
\item \code{jin_L} - for the \code{"bootDependent"} test, the automatically selected lag.
\item \code{bart_L} - for the \code{"bootBartlett"} test, the automatically selected lag.
\item \code{jin_cov} - for the \code{"bootDependent"} test, the bootstrapped asymptotic covariance matrix.
\item \code{bart_cov} - for the \code{"bootBartlett"} test, the bootstrapped asymptotic covariance matrix.
\item \code{jin_pval} - for the \code{"bootDependent"} test, the pseudo p-value.
\item \code{bart_pval} - for the \code{"bootBartlett"} test, the pseudo p-value.
}
}
\description{
Perform a hypothesis test for equality of autocovariance functions for two time series with one or more of the following methods: (Weighted) Independent, (Weighted) Dependent, Bootstrapped Dependent and Bootstrapped Bartlett. The former two tests assess equality up to a fixed lag, while the latter two select the "optimal lag" for testing the hypothesis using an AIC-like penalty at each lag. The tests can handle multivariate time series, but the computations become considerably more intense with an increase in dimension.
}
\details{
Consider two \eqn{m}-dimensional, stationary time series with mean zero. \code{autocovarianceTest} tests for equality of autocovariance functions of the respective series. The \code{"Independent"} test is given by Lund et. al (2009). Their test assumes
independence of the two series and quantifies the distribution of the autocovariance differences through Bartlett's formula (see Brockwell and Davids (1991)). \code{autocovarianceTest} provides the corresponding asymptotic covariance matrix under the null, \code{dep_cov}, the
chi-square test statistic, \code{ind_stat} and its associated degrees of freedom and p-value, \code{ind_df} and \code{ind_pval}. The \code{"Dependent"} test is an extension of the \code{"Independent"} test to linearly dependent series. If the independence of the series is
not obvious, the \code{"Dependent"} test can be used without penalty. Both tests assess equality of autocovariances up to a fixed lag \code{L}.

When running either the \code{"Independent"} or \code{"Dependent"} test, \code{autocovarianceTest} also computes weighted variants. The weighting idea is borrowed from goodness-of-fit testing (see Fisher and Gallagher (2012)). We assign to lags \eqn{0, 1, \dots , L} weights
\eqn{1, L/(L + 1), \dots , 1/(L + 1)}. The corresponding test statistics have distributions that can be approximated by a Gamma distribution. For the \code{"Independent"} test, \code{autocovarianceTest} outputs the test statistic, \code{ind_weight_stat}, its gamma distribtion
parameters, \code{ind_alpha} and \code{ind_beta}, and the associated p-value \code{ind_weight_pval}. The same statistics for \code{"Dependent"} are given by \code{dep_weight_stat}, \code{dep_alpha}, \code{dep_beta} and \code{dep_weight_pval}. All previously mentioned tests
require Gaussianity of the series. The following two tests can be applied to series with non-normal distributions.

The \code{"bootDependent"} (see Jin et. al (2019)) and \code{"bootBartlett"} tests compute a similar test statistic to the \code{"Dependent"} test but through a block of blocks bootstrap. The original test statistic is monotone increasing with lag, so the tests apply
an AIC-like penalty to choose the "best" lag (up to \code{L}) to perform the hypothesis test with. This provides some desirable properties under the alternative hypothesis. \code{prewhiten = TRUE} fits a large VAR(p) process to \code{X} and \code{Y} and then applies
the tests to the residuals of said process. It is highly recommended as prewhitening improves empirical rejection rates. \code{B} and \code{b} control the number of bootstrap resamples and block length used in the \code{"bootDependent"} and \code{"bootBartlett"} tests. \code{"bootBartlett"} replaces the Jin
et. al (2019) covariance estimate with the \code{"Dependent"} estimate averaged over each bootstrap resample. For the \code{"bootDependent"} test, \code{autocovarianceTest} outputs the test statistic, \code{jin_stat}, the optimally selected lag, \code{jin_L}, the
estimated covariance under the null, \code{jin_cov}, and psuedo p-value \code{jin_pval}. The same statistics for \code{"bootBartlett"} are given by \code{bart_stat}, \code{bart_L}, \code{bart_cov} and \code{bart_pval}.

For more on key assumptions behind each of the tests and a thorough example, please see the vignette.
}
\examples{
## Simulate two marignal AR(1) processes that are correlated at lag zero and run tests
set.seed(12345)

# Simulate correlated errors
Sigma <- matrix(c(1, 0.8, 0.8, 1), 2, 2) # covariance of errors
n <- 1000 # sample size
Errors <- mvtnorm::rmvnorm(n, mean = c(0, 0), sigma = Sigma)

# Simultate correlated AR(1) processes under the null hypothesis
X <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 1])) 
Y <- matrix(arima.sim(model = list(ar  = 0.8), n, innov = Errors[ , 2]))

# Run tests for equality of autocovariance functions up to lag 5
autocovarianceTest(X, Y, L = 5, test = c("Independent","Dependent","bootDependent","bootBartlett"))
# All tests fail to reject the null

## In males and females, do monthly lung deaths in the UK from 1974-1979
## have the same autocovariance structure?
## Here we carefully go through the process of making sure the series 
## adhere to the test assumptions
?UKLungDeaths 

n <- length(mdeaths) # series length
# plot series
plot(mdeaths)
plot(fdeaths)
# Obvious non-zero means and not stationary 

# Standardize by monthly means and standard deviations to ensure stationarity
male_matrix <- matrix(mdeaths, 6, 12, byrow = TRUE)
female_matrix <- matrix(fdeaths, 6, 12, byrow = TRUE)
# standardize series
male_stnd <- c(t(apply(male_matrix, 2, scale)))
female_stnd <- c(t(apply(female_matrix, 2, scale)))

# Plot again to check stationarity
t <- 1:n
plot(t, male_stnd, type = "l") # definitely has a trend
plot(t, female_stnd, type = "l") # maybe a slight trend

# Detrend both series to adhere to mean zero assumption + stationarity
male_stnd <- lm(male_stnd ~ t)$residuals
female_stnd <- lm(female_stnd ~ t)$residuals
plot(t, male_stnd, type = "l") # reasonably stationary, 
plot(t, female_stnd, type = "l") # reasonably stationary

# Both series are reasonably Gaussian
qqnorm(male_stnd)
qqline(male_stnd)
qqnorm(female_stnd)
qqline(female_stnd)

# Check for correlation
ccf(female_stnd, male_stnd) # Clear lag zero correlation

# The series are likely gaussian and dependent, hence it makes most sense to use "Dependent" test
autocovarianceTest(matrix(male_stnd), matrix(female_stnd), 5, test = "Dependent")

# Both the weighted and unweighted tests fail to reject the null hypothesis
# The weighted test comes close to rejecting, which makes sense given the large difference in lag 0
acf(cbind(male_stnd, female_stnd), type = "covariance", lag.max = 5)

}
\references{
Brockwell, P. J., and Davis, R. A. (1991). \emph{Time series: theory and methods (Second ed.). New York: Springer-Verlag. }\cr
\cr Fisher, T. J., and Gallagher, C. M. (2012).  \emph{New weighted portmanteau statistics for time series goodness  of  fit  testing, J.  Amer.  Statist.  Assoc., 107(498),  777–787}\cr
\cr Jin, L., Cai, L., and Wang, S. (2019).  \emph{A computational bootstrap procedure to compare two dependent  time  series, Journal  of  Statistical  Computation  and  Simulation, 89(15), 2831–2847}\cr
\cr Lund, R., Bassily, H., and Vidakovic, B. (2009). \emph{Testing equality of stationary autocovariances, Journal of Time Series Analysis, 30(3), 332–348}\cr
}
