% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\name{autocovarianceTest}
\alias{autocovarianceTest}
\title{Title}
\usage{
autocovarianceTest(
  X,
  Y,
  L,
  test = "Dependent",
  alpha = 0.05,
  B = 500,
  prewhiten = FALSE
)
}
\arguments{
\item{X}{}

\item{Y}{}

\item{L}{}

\item{test}{}

\item{alpha}{}

\item{B}{}

\item{prewhiten}{}
}
\value{

}
\description{
Title
}
