% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\docType{data}
\name{cityTemps}
\alias{cityTemps}
\title{Monthly High Temperatures in New York and Cleveland from 1960 to 2019}
\format{
An object of class \code{"data.frame"} with 1440 rows and 4 columns.
\describe{
\item{City}{the city, CLE or NYC.}
\item{Year}{the year of the high temperature.}
\item{Month }{the month of the high temperature.}
\item{TMAX}{the high temperature (degrees Farenheit).}
}
}
\source{
https://www.ncdc.noaa.gov/cdo-web/datatools/findstation
}
\usage{
data(cityTemps)
}
\description{
Monthly high temperatures in New York City and Cleveland from 1960 to 2019. The stations that record the data are located at Cleveland Hopkins International Airport and John F. Kennedy International Airport.
}
\keyword{datasets}
