% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocovarianceTestingWrappers.R
\docType{data}
\name{cityTemps}
\alias{cityTemps}
\title{Monthly High Temperatures in New York and Cleveland from 1960 to 2019}
\format{
An object of class \code{"data.frame"} with 1440 rows and 4 columns.
\describe{
\item{City}{city, CLE or NYC}
\item{Year}{year}
\item{Month }{month}
\item{TMAX}{high temperature (degrees Farenheit)}
}
}
\source{
https://www.ncdc.noaa.gov/cdo-web/datatools/findstation
}
\usage{
data(cityTemps)
}
\description{
Monthly High Temperatures in New York and Cleveland from 1960 to 2019
}
\keyword{datasets}
